@@ -9,21 +9,28 @@ import org.springframework.http.HttpMethod;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
 import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.access.AccessDeniedHandler;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
 import static com.blogsystem.security.constants.SecurityConstants.SYSTEM_WHITELIST;
 
 @Configuration
 @EnableWebSecurity
-@EnableGlobalMethodSecurity(prePostEnabled = true)
+@EnableMethodSecurity
 @RequiredArgsConstructor
 public class WebSecurityConfig {
     private final JwtAuthorizationFilter jwtAuthorizationFilter;
     private final AuthenticationLoggingFilter authenticationLoggingFilter;
+    private final AuthenticationEntryPoint unauthorizedHandler;
+    private final AccessDeniedHandler accessDeniedHandler;
+
     @Bean
     public AuthenticationManager authenticationManager(
             AuthenticationConfiguration authenticationConfiguration) throws Exception {
@@ -32,21 +39,27 @@ public class WebSecurityConfig {
 
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http.csrf().disable()
+        http.csrf(AbstractHttpConfigurer::disable)
                 // disabling csrf since we won't use form login
                 // giving permission to every request for /login endpoint
-                .authorizeRequests()
-                .antMatchers(HttpMethod.POST, SYSTEM_WHITELIST).permitAll()
-                // for everything else, the user has to be authenticated
-                .anyRequest().authenticated()
-                .and()
+                .authorizeHttpRequests(
+                        request -> request.requestMatchers(HttpMethod.POST, SYSTEM_WHITELIST)
+                                .permitAll()
+                                // for everything else, the user has to be authenticated
+                                .anyRequest()
+                                .authenticated()
+                )
+                // setting custom entry point for unauthenticated request
+                // setting custom access denied handler for not authorized request
+                .exceptionHandling(handler -> handler
+                        .accessDeniedHandler(accessDeniedHandler)
+                        .authenticationEntryPoint(unauthorizedHandler))
                 // adding the custom filter before UsernamePasswordAuthenticationFilter in the filter chain
                 .addFilterBefore(jwtAuthorizationFilter, UsernamePasswordAuthenticationFilter.class)
-                .addFilterAfter(authenticationLoggingFilter,UsernamePasswordAuthenticationFilter.class)
+                .addFilterAfter(authenticationLoggingFilter, UsernamePasswordAuthenticationFilter.class)
                 // setting stateless session, because we choose to implement Rest API
-                .sessionManagement()
-                .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-                .and();
+                .sessionManagement(s -> s
+                        .sessionCreationPolicy(SessionCreationPolicy.STATELESS));
 
         return http.build();
     }
