@@ -1,11 +1,12 @@
 package com.blogsystem.controller.auth;
 
+import com.blogsystem.common.ResponseBody;
 import com.blogsystem.dto.auth.LoginRequest;
 import com.blogsystem.dto.auth.LoginResponse;
-import com.blogsystem.security.service.impl.AuthServiceImpl;
+import com.blogsystem.dto.auth.RegisterAccountResponse;
+import com.blogsystem.service.AuthService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,15 +17,16 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/auth")
 @RequiredArgsConstructor
 public class AuthController {
-    private final AuthServiceImpl authServiceImpl;
+    private final AuthService authService;
+
     @PostMapping("/login")
-    public ResponseEntity<LoginResponse> login(@RequestBody LoginRequest loginRequest) {
-        var token = authServiceImpl.login(loginRequest);
-        return new ResponseEntity<>(token, HttpStatus.OK);
+    public ResponseBody<LoginResponse> login(@RequestBody LoginRequest loginRequest) {
+        var token = authService.login(loginRequest);
+        return new ResponseBody<>(HttpStatus.OK, token);
     }
-    @PostMapping("/logout")
-    public ResponseEntity<Void> logout() {
-//        authService.removeToken();
-        return new ResponseEntity<>(HttpStatus.OK);
+
+    @PostMapping("/register")
+    public ResponseBody<RegisterAccountResponse> logout() {
+        return new ResponseBody<>(HttpStatus.OK, null, "Success");
     }
 }
