@@ -1,11 +1,11 @@
 package com.blogsystem.entity;
 
 import com.blogsystem.enums.RoleType;
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.annotations.GenericGenerator;
 
-import javax.persistence.*;
 import java.time.OffsetDateTime;
 import java.util.UUID;
 @Entity
